Move connection config into config.go
Relocate connection-specific configuration types and helpers from
conn.go so connection behavior is easier to scan.

Keep the moved declarations unchanged.
